refactor: 데이터 클래스 presentation으로 이동
