+++ presentation/src/main/java/com/cheocharm/presentation/model/GroupMember.kt
@@ -1,3 +1,3 @@
-package com.cheocharm.domain
+package com.cheocharm.presentation.model
 
 data class GroupMember(val imageUrl: String? = null)
